requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==3.13.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-redis==5.4.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.8
mysqlclient==2.1.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.2
gunicorn